requirements.txt
django==3.2
django-cors-headers 
djangorestframework
gunicorn
psycopg2-binary==2.9.5
python-keycloak==2.15.3
requests==2.28.2